+++ addons/binding/org.openhab.binding.avmfritz/src/main/java/org/openhab/binding/avmfritz/internal/config/AvmFritzConfiguration.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.avmfritz.config;
+package org.openhab.binding.avmfritz.internal.config;
 
 import org.apache.commons.lang.builder.ToStringBuilder;
